@@ -24,8 +24,8 @@
 		"eslint-plugin-svelte": "^2.30.0",
 		"prettier": "^2.8.0",
 		"prettier-plugin-svelte": "^2.10.1",
-		"svelte": "^4.0.0",
-		"svelte-check": "^3.4.3",
+		"svelte": "^5.19.4",
+		"svelte-check": "^4.1.4",
 		"sveltekit-adapter-deno": "^0.9.2",
 		"tslib": "^2.4.1",
 		"typescript": "^5.0.0",
