@@ -1,6 +1,7 @@
 import type { Subject } from '$lib/weather_data/Subject.interface';
 
 export interface Observer<T = any> {
-	subject?: Subject;
+	id?: string;
+	subject?: Subject | null;
 	notify(data: T): void;
 }
